Arreglados varios bugs

+++ inc/utils.h
@@ -6,7 +6,7 @@
 /*   By: vzurera- <vzurera-@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/12 12:21:57 by vzurera-          #+#    #+#             */
-/*   Updated: 2026/03/14 19:09:46 by vzurera-         ###   ########.fr       */
+/*   Updated: 2026/03/16 11:54:36 by vzurera-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,10 +21,10 @@
 #pragma region "Methods"
 
 	size_t	ft_strlen(const char *str);
-	int		ft_strlcpy(char *dst, const char *src, size_t dstsize);
+	size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize);
 	char	*ft_strchr(const char *s, int c);
 	int		ft_strcmp(const char *s1, const char *s2);
-	int		ft_strncmp(const char *s1, const char *s2, int n);
+	int		ft_strncmp(const char *s1, const char *s2, size_t n);
 	void	*ft_memset(void *b, int c, int len);
 	int		ft_isspace(int c);
 	int		ft_atos(const char *str, size_t *result);
